+++ src/azure-firewall/HISTORY.rst
@@ -1,7 +1,19 @@
 .. :changelog:
 
- Release History
+Release History
 ===============
+
+0.1.6
+++++++
+* `az network firewall ip-config`: bug fix to support multiple ip-config creation.
+
+0.1.5
+++++++
+* `az network firewall application-rule`: bug fix.
+
+0.1.4
+++++++
+* `az network firewall policy`: support firewall policy and its child resource.
 
 0.1.3
 ++++++